Move QueryProvider to main layout
