+++ components/QueryProvider.tsx
@@ -1,0 +1,32 @@
+"use client";
+
+import {
+  QueryClient,
+  QueryClientProvider,
+  keepPreviousData,
+} from "@tanstack/react-query";
+import { useMemo } from "react";
+
+export default function QueryProvider({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  const queryClient = useMemo(
+    () =>
+      new QueryClient({
+        defaultOptions: {
+          queries: {
+            retry: 1,
+            placeholderData: keepPreviousData,
+            refetchOnWindowFocus: false,
+          },
+        },
+      }),
+    [],
+  );
+
+  return (
+    <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+  );
+}
